chatbox/web-server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
atlassian-python-api==3.41.1
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.4.2
docopt==0.6.2
docx2txt==0.8
exceptiongroup==1.1.2
fastapi==0.99.1
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-sock==0.6.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gevent==23.9.1
google-auth==2.23.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
html2text==2020.1.16
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
pip-install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
kubernetes==28.1.0
langchain==0.0.266
langsmith==0.0.26
llama-hub==0.0.25
llama-index==0.8.8
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mongoengine==0.27.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
package-name==0.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pipreqs==0.4.13
posthog==3.0.1
protobuf==4.24.0
psutil==5.9.5
pulsar-client==3.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pyee==9.0.4
PyJWT==2.8.0
pymongo==4.4.1
pyparsing==3.0.9
pypdf==3.15.0
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.7.1
pythonenv==0.0.34
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
safetensors==0.3.2
sentencepiece==0.1.99
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
starlette==0.27.0
stripe==6.0.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.0
transformers==4.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
validators==0.20.0
watchfiles==0.19.0
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.0
